refactor(ai): split SkillManagerDialog into types+utils+sub-components
Extract types, utility functions, SkillFormFields, and SkillItemList
from 617-line SkillManagerDialog.tsx into separate files. No behavior changes.

Closes #1239
