+++ apps/desktop/renderer/src/features/ai/SkillManagerDialog.tsx
@@ -1,342 +1,25 @@
 import { useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 
-import type { IpcError, IpcResponseData } from "@shared/types/ipc-generated";
 import { SystemDialog } from "../../components/features/AiDialogs/SystemDialog";
 import { Dialog, Text } from "../../components/primitives";
 import { useConfirmDialog } from "../../hooks/useConfirmDialog";
-import { i18n } from "../../i18n";
 import { invoke } from "../../lib/ipcClient";
 import { getHumanErrorMessage } from "../../lib/errorMessages";
 import { Button } from "../../components/primitives/Button";
-import { Checkbox } from "../../components/primitives/Checkbox";
-import { Input } from "../../components/primitives/Input";
 import { Textarea } from "../../components/primitives/Textarea";
-import { Select } from "../../components/primitives/Select";
-import { Label } from "../../components/primitives/Label";
 
-export type CustomSkillListItem =
-  IpcResponseData<"skill:custom:list">["items"][number];
+import type { CustomSkillListItem, SkillFormState } from "./skill-manager.types";
+import {
+  DEFAULT_FORM,
+  buildSkillDraftFromDescription,
+  parseContextRulesText,
+  readFieldName,
+} from "./skill-manager-utils";
+import { SkillFormFields } from "./SkillFormFields";
+import { SkillItemList } from "./SkillItemList";
 
-type SkillFormState = {
-  name: string;
-  description: string;
-  promptTemplate: string;
-  inputType: "selection" | "document";
-  scope: "global" | "project";
-  enabled: boolean;
-  contextRulesText: string;
-};
-
-type CustomSkillContextRules = Record<string, string | number | boolean>;
-
-const DEFAULT_FORM: SkillFormState = {
-  name: "",
-  description: "",
-  promptTemplate: "",
-  inputType: "selection",
-  scope: "project",
-  enabled: true,
-  contextRulesText: "{}",
-};
-
-function buildSkillDraftFromDescription(
-  description: string,
-): Pick<
-  SkillFormState,
-  "name" | "description" | "promptTemplate" | "inputType" | "contextRulesText"
-> {
-  const normalized = description.trim();
-  const shortName =
-    normalized.slice(0, 16) || i18n.t("ai.skillManager.aiGeneratedSkill");
-
-  return {
-    name: shortName,
-    description: normalized,
-    promptTemplate: `请根据以下要求处理文本：${normalized}\n\n原文：{{input}}`,
-    inputType: normalized.includes("续写") ? "document" : "selection",
-    contextRulesText: JSON.stringify({ style_guide: true }, null, 2),
-  };
-}
-
-function parseContextRulesText(
-  text: string,
-):
-  | { ok: true; data: CustomSkillContextRules }
-  | { ok: false; message: string } {
-  try {
-    const parsed: unknown = JSON.parse(text);
-    if (
-      typeof parsed !== "object" ||
-      parsed === null ||
-      Array.isArray(parsed)
-    ) {
-      return {
-        ok: false,
-        message: i18n.t("ai.skillManager.contextRulesMustBeObject"),
-      };
-    }
-    return { ok: true, data: parsed as CustomSkillContextRules };
-  } catch {
-    return {
-      ok: false,
-      message: i18n.t("ai.skillManager.contextRulesInvalidJson"),
-    };
-  }
-}
-
-function readFieldName(error: IpcError): string | null {
-  const details = error.details;
-  if (!details || typeof details !== "object") {
-    return null;
-  }
-  const fieldName = (details as { fieldName?: unknown }).fieldName;
-  return typeof fieldName === "string" ? fieldName : null;
-}
-
-// ============================================================================
-// Extracted sub-components
-// ============================================================================
-
-/**
- * SkillFormFields – the form inputs for creating/editing a custom skill.
- */
-function SkillFormFields(props: {
-  heading: string;
-  form: SkillFormState;
-  onFormChange: React.Dispatch<React.SetStateAction<SkillFormState>>;
-  fieldErrors: Record<string, string>;
-}): JSX.Element {
-  const { t } = useTranslation();
-
-  return (
-    <section className="space-y-2">
-      <Text size="tiny" color="muted" className="uppercase tracking-wide">
-        {props.heading}
-      </Text>
-
-      <Label className="block text-xs text-[var(--color-fg-muted)]">
-        {t("ai.skillManager.fieldName")}
-        <Input
-          value={props.form.name}
-          onChange={(e) =>
-            props.onFormChange((prev) => ({ ...prev, name: e.target.value }))
-          }
-          className="mt-1 w-full rounded border border-[var(--color-border-default)] bg-[var(--color-bg-base)] p-2 text-sm"
-          data-testid="skill-form-name"
-        />
-        {props.fieldErrors.name && (
-          <span className="mt-1 block text-xs text-[var(--color-error)]">
-            {props.fieldErrors.name}
-          </span>
-        )}
-      </Label>
-
-      <Label className="block text-xs text-[var(--color-fg-muted)]">
-        {t("ai.skillManager.fieldDescription")}
-        <Input
-          value={props.form.description}
-          onChange={(e) =>
-            props.onFormChange((prev) => ({
-              ...prev,
-              description: e.target.value,
-            }))
-          }
-          className="mt-1 w-full rounded border border-[var(--color-border-default)] bg-[var(--color-bg-base)] p-2 text-sm"
-          data-testid="skill-form-description"
-        />
-        {props.fieldErrors.description && (
-          <span className="mt-1 block text-xs text-[var(--color-error)]">
-            {props.fieldErrors.description}
-          </span>
-        )}
-      </Label>
-
-      <Label className="block text-xs text-[var(--color-fg-muted)]">
-        {t("ai.skillManager.fieldPromptTemplate")}
-        <Textarea
-          value={props.form.promptTemplate}
-          onChange={(e) =>
-            props.onFormChange((prev) => ({
-              ...prev,
-              promptTemplate: e.target.value,
-            }))
-          }
-          placeholder={t("ai.skillManager.promptPlaceholder")}
-          className="mt-1 w-full min-h-20 rounded border border-[var(--color-border-default)] bg-[var(--color-bg-base)] p-2 text-sm"
-          data-testid="skill-form-prompt-template"
-        />
-        {props.fieldErrors.promptTemplate && (
-          <span
-            className="mt-1 block text-xs text-[var(--color-error)]"
-            data-testid="skill-form-error-promptTemplate"
-          >
-            {props.fieldErrors.promptTemplate}
-          </span>
-        )}
-      </Label>
-
-      <div className="grid grid-cols-2 gap-2">
-        <div className="block text-xs text-[var(--color-fg-muted)]">
-          <Label className="block text-xs text-[var(--color-fg-muted)] mb-1">
-            {t("ai.skillManager.fieldInputType")}
-          </Label>
-          <Select
-            value={props.form.inputType}
-            onValueChange={(val) =>
-              props.onFormChange((prev) => ({
-                ...prev,
-                inputType: val as "selection" | "document",
-              }))
-            }
-            options={[
-              {
-                value: "selection",
-                label: t("ai.skillManager.inputTypeSelection"),
-              },
-              {
-                value: "document",
-                label: t("ai.skillManager.inputTypeDocument"),
-              },
-            ]}
-            className="mt-1 w-full text-sm"
-            data-testid="skill-form-input-type"
-            fullWidth
-          />
-        </div>
-
-        <div className="block text-xs text-[var(--color-fg-muted)]">
-          <Label className="block text-xs text-[var(--color-fg-muted)] mb-1">
-            {t("ai.skillManager.fieldScope")}
-          </Label>
-          <Select
-            value={props.form.scope}
-            onValueChange={(val) =>
-              props.onFormChange((prev) => ({
-                ...prev,
-                scope: val as "global" | "project",
-              }))
-            }
-            options={[
-              { value: "project", label: t("ai.skillManager.scopeProject") },
-              { value: "global", label: t("ai.skillManager.scopeGlobal") },
-            ]}
-            className="mt-1 w-full text-sm"
-            data-testid="skill-form-scope"
-            fullWidth
-          />
-        </div>
-      </div>
-
-      <Label className="block text-xs text-[var(--color-fg-muted)]">
-        {t("ai.skillManager.fieldContextRules")}
-        <Textarea
-          value={props.form.contextRulesText}
-          onChange={(e) =>
-            props.onFormChange((prev) => ({
-              ...prev,
-              contextRulesText: e.target.value,
-            }))
-          }
-          className="mt-1 w-full min-h-16 rounded border border-[var(--color-border-default)] bg-[var(--color-bg-base)] p-2 text-sm"
-          data-testid="skill-form-context-rules"
-        />
-        {props.fieldErrors.contextRules && (
-          <span className="mt-1 block text-xs text-[var(--color-error)]">
-            {props.fieldErrors.contextRules}
-          </span>
-        )}
-      </Label>
-
-      <Checkbox
-        checked={props.form.enabled}
-        onCheckedChange={(val) =>
-          props.onFormChange((prev) => ({
-            ...prev,
-            enabled: val === true,
-          }))
-        }
-        label={t("ai.skillManager.enableSkill")}
-        data-testid="skill-form-enabled"
-      />
-    </section>
-  );
-}
-
-/**
- * SkillItemList – the existing custom skill list with edit/delete buttons.
- */
-function SkillItemList(props: {
-  items: CustomSkillListItem[];
-  loading: boolean;
-  onEdit: (item: CustomSkillListItem) => void;
-  onDelete: (item: CustomSkillListItem) => void;
-}): JSX.Element {
-  const { t } = useTranslation();
-
-  return (
-    <section className="space-y-2">
-      <Text size="tiny" color="muted" className="uppercase tracking-wide">
-        {t("ai.skillManager.customSkillList")}
-      </Text>
-      {props.loading ? (
-        <Text size="small" color="muted">
-          {t("ai.skillManager.loading")}
-        </Text>
-      ) : props.items.length === 0 ? (
-        <Text size="small" color="muted">
-          {t("ai.skillManager.noCustomSkills")}
-        </Text>
-      ) : (
-        <div className="space-y-2" data-testid="skill-manager-list">
-          {props.items.map((item) => (
-            <div
-              key={item.id}
-              className="rounded border border-[var(--color-border-default)] p-2"
-            >
-              <div className="flex items-start justify-between gap-2">
-                <div>
-                  <Text size="small" weight="semibold">
-                    {item.name}
-                  </Text>
-                  <Text size="tiny" color="muted">
-                    {item.scope === "project"
-                      ? t("ai.skillManager.scopeProject")
-                      : t("ai.skillManager.scopeGlobal")}{" "}
-                    · {item.inputType}
-                  </Text>
-                </div>
-                <div className="flex items-center gap-2">
-                  <Button
-                    type="button"
-                    className="px-2 py-1 rounded border border-[var(--color-border-default)] text-xs"
-                    onClick={() => props.onEdit(item)}
-                    data-testid={`skill-item-edit-${item.id}`}
-                  >
-                    {t("ai.skillManager.edit")}
-                  </Button>
-                  <Button
-                    type="button"
-                    className="px-2 py-1 rounded border border-[var(--color-error)]/30 text-xs text-[var(--color-error)]"
-                    onClick={() => props.onDelete(item)}
-                    data-testid={`skill-item-delete-${item.id}`}
-                  >
-                    {t("ai.skillManager.delete")}
-                  </Button>
-                </div>
-              </div>
-            </div>
-          ))}
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ============================================================================
-// Main component
-// ============================================================================
-
+export type { CustomSkillListItem } from "./skill-manager.types";
 /**
  * SkillManagerDialog provides custom skill CRUD + AI-assisted drafting.
  *
