+++ apps/desktop/renderer/src/features/ai/skill-manager.types.ts
@@ -1,0 +1,16 @@
+import type { IpcResponseData } from "@shared/types/ipc-generated";
+
+export type CustomSkillListItem =
+  IpcResponseData<"skill:custom:list">["items"][number];
+
+export type SkillFormState = {
+  name: string;
+  description: string;
+  promptTemplate: string;
+  inputType: "selection" | "document";
+  scope: "global" | "project";
+  enabled: boolean;
+  contextRulesText: string;
+};
+
+export type CustomSkillContextRules = Record<string, string | number | boolean>;
